create CalendarComponent.jsx
