+++ app/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^18.1.2",
+    "@expo/ngrok": "^4.1.0",
     "@react-native-community/cli": "^11.3.7",
     "expo": "^48.0.17",
     "expo-status-bar": "~1.4.4",
@@ -19,7 +20,8 @@
     "react-dom": "^18.2.0",
     "react-native": "0.71.14",
     "react-native-calendars": "1.1284.0",
-    "react-native-web": "~0.18.10"
+    "react-native-web": "~0.18.10",
+    "react-native-simple-dialogs": "2.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
